+++ src/main/java/com/hongikbros/jobmanager/notice/domain/notice/NoticeRepository.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.notice.notice.domain;
+package com.hongikbros.jobmanager.notice.domain.notice;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
