+++ src/main/java/com/hongikbros/jobmanager/skill/domain/skill/SkillRepository.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.skill.domain;
+package com.hongikbros.jobmanager.skill.domain.skill;
 
 import java.util.List;
