+++ src/test/java/com/hongikbros/jobmanager/notice/notice/ui/NoticeViewServiceTest.java
@@ -18,18 +18,20 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.hongikbros.jobmanager.common.domain.Association;
 import com.hongikbros.jobmanager.common.fixture.sessionmember.SessionMemberFixture;
 import com.hongikbros.jobmanager.common.utils.TestObjectUtils;
-import com.hongikbros.jobmanager.notice.bookmark.BookmarkRepository;
-import com.hongikbros.jobmanager.notice.company.Company;
-import com.hongikbros.jobmanager.notice.company.CompanyRepository;
-import com.hongikbros.jobmanager.notice.notice.domain.ApplyUrl;
-import com.hongikbros.jobmanager.notice.notice.domain.Duration;
-import com.hongikbros.jobmanager.notice.notice.domain.Notice;
-import com.hongikbros.jobmanager.notice.notice.domain.NoticeDescription;
-import com.hongikbros.jobmanager.notice.notice.domain.NoticeRepository;
-import com.hongikbros.jobmanager.skill.domain.Skill;
-import com.hongikbros.jobmanager.skill.domain.SkillNotice;
-import com.hongikbros.jobmanager.skill.domain.SkillNoticeRepository;
-import com.hongikbros.jobmanager.skill.domain.SkillRepository;
+import com.hongikbros.jobmanager.notice.domain.bookmark.BookmarkRepository;
+import com.hongikbros.jobmanager.notice.domain.company.Company;
+import com.hongikbros.jobmanager.notice.domain.company.CompanyRepository;
+import com.hongikbros.jobmanager.notice.domain.notice.ApplyUrl;
+import com.hongikbros.jobmanager.notice.domain.notice.Duration;
+import com.hongikbros.jobmanager.notice.domain.notice.Notice;
+import com.hongikbros.jobmanager.notice.domain.notice.NoticeDescription;
+import com.hongikbros.jobmanager.notice.domain.notice.NoticeRepository;
+import com.hongikbros.jobmanager.notice.ui.NoticeResponse;
+import com.hongikbros.jobmanager.notice.ui.NoticeViewService;
+import com.hongikbros.jobmanager.skill.domain.skill.Skill;
+import com.hongikbros.jobmanager.skill.domain.skill.SkillRepository;
+import com.hongikbros.jobmanager.skill.domain.skillnotice.SkillNotice;
+import com.hongikbros.jobmanager.skill.domain.skillnotice.SkillNoticeRepository;
 
 @ExtendWith(MockitoExtension.class)
 class NoticeViewServiceTest {
